Modification is from Procedural to Object-Oriented
